+++ problems/0022/compute.jl
@@ -1,4 +1,4 @@
-function compute(file::String)::Int64
+function compute(file::String)::Int
     chars = Dict(Char(Int('A') + i) => i + 1 for i ∈ 0:25)
     return sum(i * chars[c] for (i, name) ∈ enumerate(sort(split(replace(file, "\"" => ""), ","))) for c ∈ name)
 end
