+++ problems/0031/compute.jl
@@ -1,4 +1,4 @@
-function compute(coins::Array{Int64,1}, n::Int64)::Int64
+function compute(coins::Array{Int,1}, n::Int)::Int
     combinations = vcat([1], repeat([0], n))
     for i ∈ 1:length(coins)
         for j ∈ coins[i] + 1:n + 1
