+++ problems/0071/compute.jl
@@ -1,1 +1,1 @@
-compute(n::Int64)::Int64 = (n - 5) ÷ 7 * 3 + 2
+compute(n::Int)::Int = (n - 5) ÷ 7 * 3 + 2
