+++ problems/0071/solution.jl
@@ -1,7 +1,7 @@
 using BenchmarkTools
 BenchmarkTools.DEFAULT_PARAMETERS.samples = 100
 
-compute(n::Int64)::Int64 = (n - 5) ÷ 7 * 3 + 2
+compute(n::Int)::Int = (n - 5) ÷ 7 * 3 + 2
 
 compute(8)
