+++ problems/0493/compute.jl
@@ -1,1 +1,1 @@
-compute(n::Int64, k::Int64)::Float64 = round(7(1 - prod(1 .- k ./ (6n ÷ 7 + 1:n))), digits=9)
+compute(n::Int, k::Int)::Float64 = round(7(1 - prod(1 .- k ./ (6n ÷ 7 + 1:n))), digits=9)
